refactor - organize files into folders
