+++ jobs/Backend/Task/ExchangeRateUpdater/Domain/Currency.cs
@@ -1,4 +1,4 @@
-﻿namespace ExchangeRateUpdater
+﻿namespace ExchangeRateUpdater.Domain
 {
     public class Currency
     {
